监控代码

+++ src/main/java/com/zgiot/app/server/module/sfmonitor/controller/SFMonitorSignalWrapperController.java
@@ -1,9 +1,12 @@
 package com.zgiot.app.server.module.sfmonitor.controller;
 
 import com.alibaba.fastjson.JSON;
+import com.github.pagehelper.PageHelper;
 import com.zgiot.app.server.module.metrictag.pojo.MetricTag;
+import com.zgiot.app.server.module.sfmonitor.mapper.SFMonDisplayZoneMapper;
 import com.zgiot.app.server.module.util.validate.ControllerUtil;
 import com.zgiot.app.server.service.impl.mapper.MetricTagMapper;
+import com.zgiot.app.server.service.impl.mapper.MetricTagRelationMapper;
 import com.zgiot.common.constants.GlobalConstants;
 import com.zgiot.common.restcontroller.ServerResponse;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,26 +22,36 @@ import java.util.Date;
 public class SFMonitorSignalWrapperController {
     @Autowired
     private MetricTagMapper metricTagMapper;
+    @Autowired
+    private MetricTagRelationMapper metricTagRelationMapper;
 
-    @GetMapping("")
-    public ResponseEntity<String> getSignalWrapper(){
+    @Autowired
+    private SFMonDisplayZoneMapper sfMonDisplayZoneMapper;
+
+    @GetMapping("/pageNum/{pageNum}/pageSize/{pageSize}")
+    public ResponseEntity<String> getSignalWrapper(@PathVariable int pageNum,@PathVariable int pageSize){
         MetricTag metricTag = new MetricTag();
+        PageHelper.startPage(pageNum,pageSize);
         return new ResponseEntity<>(ServerResponse.buildOkJson(metricTagMapper.findMetricTag(metricTag)), HttpStatus.OK);
     }
 
     @RequestMapping(value = "",method = RequestMethod.POST)
     public ResponseEntity<String> addOrEditSignalWrapper(@RequestBody String bodyStr){
         ControllerUtil.validateBodyRequired(bodyStr);
-        MetricTag metricTag = JSON.parseObject(bodyStr,MetricTag.class);
+        SignalWrapperReq signalWrapperReq = JSON.parseObject(bodyStr,SignalWrapperReq.class);
+        MetricTag metricTag = signalWrapperReq.getMetricTag();
+        int zoneId = signalWrapperReq.getZoneId();
         String name = metricTag.getTagName();
         MetricTag tag = metricTagMapper.getMetricTagByName(name);
         boolean isExist = false;
-        if(metricTag.getMetricTagId() == 0){//add
+        if(metricTag.getMetricTagId() == null){//add
             if(tag != null){
                 isExist = true;
             }else{
                 metricTag.setCreateDate(new Date());
                 metricTagMapper.addMetricTag(metricTag);
+                MetricTag metricTagTemp = metricTagMapper.getMetricTagByName(metricTag.getTagName());
+                sfMonDisplayZoneMapper.addRelSFMonTagDisplayZone(metricTagTemp.getCode(),zoneId);
             }
         }else{//edit
             if(tag != null && tag.getMetricTagId() != metricTag.getMetricTagId()){
@@ -46,6 +59,9 @@ public class SFMonitorSignalWrapperController {
             }else{
                 metricTag.setUpdateDate(new Date());
                 metricTagMapper.updateMetricTag(metricTag);
+                MetricTag metricTagTemp = metricTagMapper.getMetricTagByName(metricTag.getTagName());
+                sfMonDisplayZoneMapper.deleteRelSFMonTagDisplayZone(metricTagTemp.getCode());
+                sfMonDisplayZoneMapper.addRelSFMonTagDisplayZone(metricTagTemp.getCode(),zoneId);
             }
         }
         return new ResponseEntity<>(ServerResponse.buildOkJson(isExist), HttpStatus.OK);
@@ -53,12 +69,20 @@ public class SFMonitorSignalWrapperController {
 
     @RequestMapping(value = "/{id}",method = RequestMethod.DELETE)
     public ResponseEntity<String> deleteSignalWrapper(@PathVariable int id){
-        int signalCount = 0;
-
-        metricTagMapper.deleteThingTag(id);
-        return new ResponseEntity<>(ServerResponse.buildOkJson(null), HttpStatus.OK);
+        Integer signalCount = metricTagRelationMapper.getMetricCount(id);
+        if(signalCount == null){
+            metricTagMapper.deleteThingTag(id);
+        }
+        return new ResponseEntity<>(ServerResponse.buildOkJson(signalCount), HttpStatus.OK);
     }
 
+    @GetMapping("/{name}")
+    public ResponseEntity<String> getSignalWrapper(@PathVariable String name){
+        return new ResponseEntity<>(ServerResponse.buildOkJson(metricTagMapper.getMetricTag(name)), HttpStatus.OK);
+    }
 
-
+    @GetMapping("/zone")
+    public ResponseEntity<String> getDisplayZone(){
+        return new ResponseEntity<>(ServerResponse.buildOkJson(sfMonDisplayZoneMapper.getDisplayZone()), HttpStatus.OK);
+    }
 }
